+++ package.json
@@ -14,6 +14,7 @@
     "email-templates": "^3.1.7",
     "nodemailer": "^4.4.0",
     "passport": "^0.4.0",
+    "passport-google-oauth": "^1.0.0",
     "passport-local": "^1.0.0",
     "pug": "^2.0.0-rc.4",
     "require-fields": "^1.0.0"
