+++ src/constants/fields.js
@@ -5,5 +5,7 @@ module.exports = {
   CONFIRM_EMAIL_TOKEN    : "confirmEmailToken",
   EMAIL_CONFIRMED        : "emailConfirmed",
   RESET_PASSWORD_TOKEN   : "resetPasswordToken",
-  RESET_PASSWORD_EXPIRES : "resetPasswordExpires"
+  RESET_PASSWORD_EXPIRES : "resetPasswordExpires",
+
+  OAUTH2_TOKEN : "oauth2Token"
 }
