--- src/controllers/middleware/requireLoggedIn.js
@@ -1,7 +1,0 @@
-"use strict"
-
-module.exports = function(req, res, next) {
-  if (!req.isAuthenticated())
-    return res.sendStatus(401)
-  next();
-}
